==> notebooks/1_input_prep/4_EPOC_create_input_files/1_make_demographic_csv_files.ipynb <==
import os
import glob
import fnmatch
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from src.configs import configs

epoc = pd.read_spss(f"{configs.PROJECT_ROOT}/EPOC/EPOC_data_forSP9_mLPP.sav")

epoc.shape

#### Check variable values

epoc['Status'].unique()

epoc['Ausschluss'].unique()

epoc['Response'].value_counts(dropna=False)

epoc['Remission'].value_counts(dropna=False)

#### Delete healthy controls

target_to_list_status = ['Index']

epoc = epoc[epoc['Status'].isin(target_to_list_status)]
print(epoc.shape)

#### Delete "Ausschluss" subjects

target_to_list = ['Kein Ausschluss']

epoc = epoc[epoc['Ausschluss'].isin(target_to_list)]
print(epoc.shape)

#### Delete subjects with missing "Response" or "Remission" values

epoc = epoc.dropna(subset=['Response'])
print(epoc.shape)

epoc = epoc.dropna(subset=['Remission'])
print(epoc.shape)

Subjects with missing Response values are identical to subjects with missing Remission values

#### Check for NaNs

epoc['Geschlecht'].value_counts(dropna=False)

epoc['Alter'].value_counts(dropna=False)

#### Save files for all patients (that were not excluded from the study)

# all variables
epoc.to_csv(f'{configs.PROJECT_ROOT}/EPOC/csv_files/EPOC_all_pat_119.csv', index=False)

#### Save files for patients that have structural MRI scans

pat_labels_julia = epoc['Code'].unique()
pat_labels_julia.sort()
print(pat_labels_julia)
print(pat_labels_julia.shape)

pat_labels_julia = np.ndarray.tolist(pat_labels_julia)

pat_labels_christian = fnmatch.filter(os.listdir(f'{configs.PROJECT_ROOT}/EPOC/EPOC_BIDS/EPOC_derivatives'), 'sub-epocp*')
pat_labels_christian.sort()
pat_labels_christian

pat_labels_christian_new = []
for subj in pat_labels_christian:
    
    subj_new = subj[4:8] + '_' + subj[8] + '_' + subj[9:None]
    pat_labels_christian_new.append(subj_new)

len(pat_labels_christian_new)

epoc_T1only = epoc.loc[epoc['Code'].isin(pat_labels_christian_new)]

epoc_T1only.shape

epoc_T1only

epoc_T1only.head()

# all variables
epoc_T1only.to_csv(f'{configs.PROJECT_ROOT}/EPOC/csv_files/EPOC_T1_pat_90.csv', index=False)

#### Save file for patients that have rs-fMRI scans

pat_labels_fMRI = fnmatch.filter(os.listdir(f'{configs.PROJECT_ROOT}/EPOC/halfpipe_patients_only'), 'sub-epocp*')
pat_labels_fMRI.sort()
pat_labels_fMRI

pat_labels_fMRI_new = []
for subj in pat_labels_fMRI:
    
    subj_new = subj[4:8] + '_' + subj[8] + '_' + subj[9:None]
    pat_labels_fMRI_new.append(subj_new)

len(pat_labels_fMRI_new)

epoc_fMRI = epoc.loc[epoc['Code'].isin(pat_labels_fMRI_new)]

epoc_fMRI.shape

epoc_fMRI.head()

epoc_fMRI.to_csv(f'{configs.PROJECT_ROOT}/EPOC/csv_files/EPOC_fMRI_pat_80.csv', index=False)
