==> notebooks/1_input_prep/4_EPOC_create_input_files/5_make_demografic_sMRI_csv_file.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import pyreadstat
import seaborn as sns
from scipy import stats

from src.configs import configs

cmap = plt.cm.rainbow
import h5py

#### Select only subjects with demografic info

# select only subjects with demografic info
epoc_dem = pd.read_csv(f'{configs.PROJECT_ROOT}/EPOC/csv_files/EPOC_T1_pat_90.csv', sep=",")
pat_labels_epoc = epoc_dem['Code'].unique()
pat_labels_epoc.sort()
print(pat_labels_epoc.shape)
pat_labels_epoc = np.ndarray.tolist(pat_labels_epoc)

def create_MRI_csv(file_path, pat_labels_epoc):
    
    df = pd.read_csv(file_path, sep="\t")
    
    df.iloc[:,0] = df.iloc[:,0].str[4:8] + '_' + df.iloc[:,0].str[8] + '_' + df.iloc[:,0].str[9:13]
    
    df_dem_pat = df.loc[df.iloc[:,0].isin(pat_labels_epoc)]
    print(df_dem_pat.shape)
    
    return df_dem_pat

#### Make dataframes for freesurfer outputs

file_path = f"{configs.OUTPUT_ROOT}/3_output/EPOC/aseg_stats.txt"
df_aseg = create_MRI_csv(file_path, pat_labels_epoc)
df_aseg.head()

file_path = f"{configs.OUTPUT_ROOT}/3_output/EPOC/lh.thickness.txt"
df_lh_thick = create_MRI_csv(file_path, pat_labels_epoc)
df_lh_thick.head()

file_path = f"{configs.OUTPUT_ROOT}/3_output/EPOC/rh.thickness.txt"
df_rh_thick = create_MRI_csv(file_path, pat_labels_epoc)
df_rh_thick.head()

file_path = f"{configs.OUTPUT_ROOT}/3_output/EPOC/lh.volume.txt"
df_lh_vol = create_MRI_csv(file_path, pat_labels_epoc)
df_lh_vol.head()

file_path = f"{configs.OUTPUT_ROOT}/3_output/EPOC/rh.volume.txt"
df_rh_vol = create_MRI_csv(file_path, pat_labels_epoc)
df_rh_vol.head()

df_FS_all = pd.concat([df_aseg.iloc[:, 1:], df_lh_thick.iloc[:, 1:], df_rh_thick.iloc[:, 1:], df_lh_vol.iloc[:, 1:], df_rh_vol.iloc[:, 1:]], axis=1)

df_FS_all.shape

df_FS_all.rename(columns={"Measure:volume": "subject"},inplace = True)

### Add demografic information

epoc_dem.head()

epoc_dem = epoc_dem.sort_values(by = 'Code')
epoc_dem

epoc_dem.index = np.arange(epoc_dem.shape[0])
epoc_dem

df_FS_all.index = np.arange(df_FS_all.shape[0])
df_FS_all

df_dem_MRI = pd.concat([epoc_dem, df_FS_all], axis=1)

df_dem_MRI 

df_dem_MRI.to_csv(f'{configs.PROJECT_ROOT}/EPOC/csv_files/EPOC_MRI_pat_90.csv', index=False)
